+++ next/src/app/[lang]/legals/page.tsx
