+++ next/src/app/[lang]/legals/terms-of-service/page.tsx
@@ -3,11 +3,11 @@ export const dynamic = "force-dynamic";
 import qs from 'qs';
 import { notFound } from 'next/navigation';
 import type { Metadata } from "next";
-import { getStrapiData, getStrapiMetadata } from "../../../utils/fetchStrapiData";
-import { blockRenderer, Block, } from "../../../utils/render"
+import { getStrapiData, getStrapiMetadata } from "../../../../utils/fetchStrapiData";
+import { blockRenderer, Block, } from "../../../../utils/render"
 
-const path = "/api/impressum";
-const queryImpressum = qs.stringify({
+const path = "/api/terms-of-service";
+const queryToS = qs.stringify({
 	populate: {
 		blocks: {
 			populate: "*",
@@ -18,7 +18,7 @@ const queryImpressum = qs.stringify({
 export async function generateMetadata(): Promise<Metadata> {
   const metadata = await getStrapiMetadata(
 	path,
-	"Impressum - J42L",
+	"Terms of Service - J42L",
 	"Junior 42 Lausanne",
   );
 
@@ -28,9 +28,9 @@ export async function generateMetadata(): Promise<Metadata> {
   };
 }
 
-export default async function Impressum() {
+export default async function ToS() {
 	try {
-		const strapiData = await getStrapiData(path, queryImpressum);
+		const strapiData = await getStrapiData(path, queryToS);
 		if (strapiData.type == "NOT_FOUND") {
 			return notFound();
 		}
@@ -46,7 +46,7 @@ export default async function Impressum() {
 			</div>
 		)
 	} catch(error) {
-		console.error(`Page Impressum. ${error}`);
+		console.error(`Page Terms of Service. ${error}`);
 		throw error;
 	}
 }
