+++ next/src/app/[lang]/services/prototype/page.tsx
@@ -1,24 +1,24 @@
 export const dynamic = "force-dynamic";
 
 import qs from 'qs';
-import { notFound } from 'next/navigation';
 import type { Metadata } from "next";
-import { getStrapiData, getStrapiMetadata } from "../../../utils/fetchStrapiData";
-import { blockRenderer, Block, } from "../../../utils/render"
+import { notFound } from 'next/navigation';
+import { getStrapiData, getStrapiMetadata } from "../../../../utils/fetchStrapiData";
+import { blockRenderer, Block, } from "../../../../utils/render"
 
-const path = "/api/web-service";
-const queryWeb = qs.stringify(
+const path = "/api/prototype-service";
+const queryPrototype = qs.stringify(
   {
     populate: {
       blocks: {
         on: {
           'layout.hero': {
-			populate: "*",
-		  },
-		  'layout.anchor-tag': {
-			populate: "*",
-		  },
-          'layout.services': {
+            populate: "*",
+          },
+          'layout.anchor-tag': {
+            populate: "*",
+          },
+          "layout.services": {
             populate: {
               servicesTitle: true,
               servicesAccordions: {
@@ -34,8 +34,8 @@ const queryWeb = qs.stringify(
             populate: "*",
           },
           'layout.footer-cta': {
-					populate: "*",
-				},
+          populate: "*",
+        },
         },
       },
     },
@@ -45,20 +45,20 @@ const queryWeb = qs.stringify(
 
 export async function generateMetadata(): Promise<Metadata> {
   const metadata = await getStrapiMetadata(
-	path,
-	"Web - J42L",
-	"Junior 42 Lausanne",
+    path,
+    "Prototype - J42L",
+    "Junior 42 Lausanne",
   );
 
   return {
-	title: metadata.title,
-	description: metadata.description,
+    title: metadata.title,
+    description: metadata.description,
   };
 }
 
-export default async function Web() {
+export default async function Prototype() {
 	try {
-		const strapiData = await getStrapiData(path, queryWeb);
+		const strapiData = await getStrapiData(path, queryPrototype);
 		if (strapiData.type == "NOT_FOUND") {
 			return notFound();
 		}
@@ -74,7 +74,7 @@ export default async function Web() {
 			</div>
 		)
 	} catch(error) {
-		console.error(`Page Web. ${error}`);
+		console.error(`Page Prototype. ${error}`);
 		throw error;
 	}
 }
