+++ next/src/app/[lang]/services/web/page.tsx
@@ -3,22 +3,22 @@ export const dynamic = "force-dynamic";
 import qs from 'qs';
 import { notFound } from 'next/navigation';
 import type { Metadata } from "next";
-import { getStrapiData, getStrapiMetadata } from "../../../utils/fetchStrapiData";
-import { blockRenderer, Block, } from "../../../utils/render"
+import { getStrapiData, getStrapiMetadata } from "../../../../utils/fetchStrapiData";
+import { blockRenderer, Block, } from "../../../../utils/render"
 
-const path = "/api/automation-service";
-const queryAutomation = qs.stringify(
+const path = "/api/web-service";
+const queryWeb = qs.stringify(
   {
     populate: {
       blocks: {
         on: {
           'layout.hero': {
-            populate: "*",
-          },
-          'layout.anchor-tag': {
-          populate: "*",
-          },
-          "layout.services": {
+			populate: "*",
+		  },
+		  'layout.anchor-tag': {
+			populate: "*",
+		  },
+          'layout.services': {
             populate: {
               servicesTitle: true,
               servicesAccordions: {
@@ -34,8 +34,8 @@ const queryAutomation = qs.stringify(
             populate: "*",
           },
           'layout.footer-cta': {
-          populate: "*",
-        },
+					populate: "*",
+				},
         },
       },
     },
@@ -45,20 +45,20 @@ const queryAutomation = qs.stringify(
 
 export async function generateMetadata(): Promise<Metadata> {
   const metadata = await getStrapiMetadata(
-    path,
-    "Automation - J42L",
-    "Junior 42 Lausanne",
+	path,
+	"Web - J42L",
+	"Junior 42 Lausanne",
   );
 
   return {
-    title: metadata.title,
-    description: metadata.description,
+	title: metadata.title,
+	description: metadata.description,
   };
 }
 
-export default async function Automation() {
+export default async function Web() {
 	try {
-		const strapiData = await getStrapiData(path, queryAutomation);
+		const strapiData = await getStrapiData(path, queryWeb);
 		if (strapiData.type == "NOT_FOUND") {
 			return notFound();
 		}
@@ -74,7 +74,7 @@ export default async function Automation() {
 			</div>
 		)
 	} catch(error) {
-		console.error(`Page Automation. ${error}`);
+		console.error(`Page Web. ${error}`);
 		throw error;
 	}
 }
