+++ next/src/utils/type.tsx
@@ -1,2 +1,6 @@
 export type ThemeColor = 'orange' | 'white' | 'black' | 'pale_orange';
 export type Mode = 'desktop' | 'mobile';
+export type Locale = "fr" | "en" | "de";
+export type LangParams = {
+  lang: Locale;
+};
